@@ -1,57 +1,59 @@
 {
   "name": "landing-example2",
-  "license": "MIT",
+  "version": "0.24.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --prod",
-    "test": "ng test --sourcemaps=false",
+    "build": "ng build",
+    "test": "ng test",
     "lint": "ng lint",
     "e2e": "ng e2e",
     "build:prod": "ng build --target production --build-optimizer --vendor-chunk",
-    "release": "standard-version && git push --follow-tags"
+    "release": "standard-version && git push --follow-tags",
+    "deploy": "copy .\\_redirects .\\dist\\LandingExample2\\ && netlify deploy --dir=dist\\LandingExample2 --prod"
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "^1.0.0-beta.2",
-    "@angular/animations": "5.2.9",
-    "@angular/cdk": "6.0.0-beta.4",
-    "@angular/common": "5.2.9",
-    "@angular/compiler": "5.2.9",
-    "@angular/core": "5.2.9",
-    "@angular/forms": "5.2.9",
-    "@angular/http": "5.2.9",
-    "@angular/material": "6.0.0-beta.4",
-    "@angular/platform-browser": "5.2.9",
-    "@angular/platform-browser-dynamic": "5.2.9",
-    "@angular/router": "5.2.9",
-    "core-js": "2.4.1",
-    "rxjs": "5.5.6",
-    "zone.js": "0.8.19",
-    "@types/file-saver": "1.3.0",
-    "@types/googlemaps": "^3.30.8",
-    "file-saver": "1.3.3",
-    "ng2-pdf-viewer": "^4.1.2"
+    "@agm/core": "^1.0.0-beta.5",
+    "@angular/animations": "^7.1.3",
+    "@angular/cdk": "^7.1.1",
+    "@angular/common": "^7.1.3",
+    "@angular/compiler": "^7.1.3",
+    "@angular/core": "^7.1.3",
+    "@angular/forms": "^7.1.3",
+    "@angular/http": "^7.1.3",
+    "@angular/material": "^7.1.1",
+    "@angular/platform-browser": "^7.1.3",
+    "@angular/platform-browser-dynamic": "^7.1.3",
+    "@angular/router": "^7.1.3",
+    "core-js": "^2.5.4",
+    "hammerjs": "^2.0.8",
+    "ng2-pdf-viewer": "^5.1.1",
+    "ng5-slider": "^1.1.11",
+    "ngx-bar-rating": "^1.1.0",
+    "rxjs": "^6.0.0",
+    "zone.js": "~0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.2",
-    "@angular/compiler-cli": "5.2.9",
-    "@angular/language-service": "5.2.9",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "~7.0.4",
+    "@angular/compiler-cli": "~7.0.0",
+    "@angular/language-service": "~7.0.0",
+    "@types/googlemaps": "^3.30.11",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "~4.5.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~3.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3",
-    "webpack-dev-server": "^3.1.14"
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.1.1"
   }
 }
