@@ -1,10 +1,10 @@
 import { Injector, Injectable } from '@angular/core';
 import { HttpEvent, HttpInterceptor, HttpHandler, HttpRequest, HttpResponse } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { AuthService } from '../providers/auth.service';
 
-import { AppConstants } from '../../app-constants';
+import { environment } from '@env/environment';
 
 @Injectable()
 export class AuthHeadersInterceptor implements HttpInterceptor {
@@ -16,7 +16,7 @@ export class AuthHeadersInterceptor implements HttpInterceptor {
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
     let authReq;
     const authService = this.inj.get(AuthService);
-    if (authService.token && req.url.includes(AppConstants.API_ENDPOINT)) {
+    if (authService.token && req.url.includes(environment.API_ENDPOINT)) {
       authReq = req.clone({ setHeaders: { Authorization: authService.token } });
     }
     return next.handle(authReq ? authReq : req);
