@@ -2,21 +2,20 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
 
 import { Auth } from '../../auth/auth.model';
-import { AppConstants } from '../../app-constants';
-import { StaticMethods } from '../../utils/static-methods';
+import { AppConstants } from '@app/app-constants';
+import { environment } from '@env/environment';
+import { StaticMethods } from '@core/static-methods';
 
-import { Observable } from 'rxjs/Observable';
-import { ErrorObservable } from 'rxjs/observable/errorObservable';
-import { catchError, tap, share, timeout, concat, timeoutWith, zip } from 'rxjs/operators';
-
-import 'rxjs/add/observable/of';
-
-import { Subject } from 'rxjs/Subject';
+import { Observable, Subject, of, throwError } from 'rxjs';
+import { catchError, tap, share, concat, timeoutWith } from 'rxjs/operators';
 
 import { ModalManager } from './modal-manager';
+import { User } from '@app/shared/models/user.model';
 
 
-@Injectable()
+@Injectable({
+  providedIn: 'root'
+})
 export class AuthService {
 
   private isTokenChecked = false;
@@ -51,11 +50,11 @@ export class AuthService {
   }
 
   login(user): Observable<Auth> {
-    return this.http.post<Auth>(`${AppConstants.API_ENDPOINT}login`, user)
+    return this.http.post<Auth>(`${environment.API_ENDPOINT}login`, user)
       .pipe(
         catchError((err, caught) => {
           this.mm.closeLoadingDialog();
-          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+          return throwError(StaticMethods.handleHttpResponseError(err));
         }),
         tap((res) => {
           res = new Auth(res);
@@ -69,11 +68,11 @@ export class AuthService {
   }
 
   signup(account): Observable<any> {
-    return this.http.post(`${AppConstants.API_ENDPOINT}users`, account)
+    return this.http.post(`${environment.API_ENDPOINT}users`, account)
       .pipe(
         catchError((err, caught) => {
           this.mm.closeLoadingDialog();
-          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+          return throwError(StaticMethods.handleHttpResponseError(err));
         })
       );
   }
@@ -86,11 +85,11 @@ export class AuthService {
     formData.set('certificado_libertad_tradicion', renter.certificado_libertad_tradicion.file, renter.certificado_libertad_tradicion.name);
     formData.set('rut', renter.rut.file, renter.rut.name);
 
-    return this.http.post(`${AppConstants.API_ENDPOINT}renters`, formData)
+    return this.http.post(`${environment.API_ENDPOINT}renters`, formData)
       .pipe(
         catchError((err, caught) => {
           this.mm.closeLoadingDialog();
-          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+          return throwError(StaticMethods.handleHttpResponseError(err));
         })
       );
   }
@@ -100,19 +99,19 @@ export class AuthService {
     this._token = sessionStorage.getItem('token');
     if (!this.token) {
       this.mm.closeLoadingDialog();
-      return ErrorObservable.create('No token');
+      return throwError('No token');
     }
 
     if (this.isTokenChecked) {
       if (this.user) {
-        return Observable.of(this.user);
+        return of(this.user);
       } else {
         this.mm.closeLoadingDialog();
-        return ErrorObservable.create('No token');
+        return throwError('No token');
       }
     }
 
-    return this.http.get<Auth>(`${AppConstants.API_ENDPOINT}check`)
+    return this.http.get<Auth>(`${environment.API_ENDPOINT}check`)
       .pipe(
         share(),
         tap((res) => {
@@ -130,7 +129,7 @@ export class AuthService {
           this._token = undefined;
           sessionStorage.removeItem('token');
           this.mm.closeLoadingDialog();
-          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+          return throwError(StaticMethods.handleHttpResponseError(err));
         }));
   }
 
@@ -141,7 +140,7 @@ export class AuthService {
       return this.check();
     }
     return this.tabSessionSbj.pipe(
-      timeoutWith(10000, Observable.of(undefined)),
+      timeoutWith(10000, of(undefined)),
       share(),
       concat(this.check())
     );
@@ -155,27 +154,27 @@ export class AuthService {
     sessionStorage.removeItem('token');
     this.logoutSbj.next();
 
-    return this.http.delete(`${AppConstants.API_ENDPOINT}sessions/logout`);
+    return this.http.delete(`${environment.API_ENDPOINT}sessions/logout`);
   }
 
   forgotPassword(user) {
-    return this.http.post(`${AppConstants.API_ENDPOINT}users/recover_password`, user, {
+    return this.http.post(`${environment.API_ENDPOINT}users/recover_password`, user, {
       responseType: 'text',
     })
       .pipe(
         catchError((err, caught) => {
           this.mm.closeLoadingDialog();
-          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+          return throwError(StaticMethods.handleHttpResponseError(err));
         })
       );
   }
 
   recoverPassword(user) {
-    return this.http.post(`${AppConstants.API_ENDPOINT}users/update_password`, user)
+    return this.http.post(`${environment.API_ENDPOINT}users/update_password`, user)
       .pipe(
         catchError((err, caught) => {
           this.mm.closeLoadingDialog();
-          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+          return throwError(StaticMethods.handleHttpResponseError(err));
         })
       );
   }
