@@ -2,16 +2,16 @@ import { Router, ActivatedRoute, ParamMap } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
 import { Validators, FormBuilder, FormGroup, AbstractControl, FormControl } from '@angular/forms';
 
-import { AuthService } from '../../../app/core/providers/auth.service';
-import { StaticMethods } from '../../../app/utils/static-methods';
-import { ModalManager } from '../../../app/core/providers/modal-manager';
-import { CustomValidators } from '../../../app/shared/custom-validators';
+import { AuthService } from '@core/providers/auth.service';
+import { StaticMethods } from '@core/static-methods';
+import { ModalManager } from '@core/providers/modal-manager';
+import { CustomValidators } from '@shared/custom-validators';
 
 import { MatSnackBar, MatSnackBarConfig } from '@angular/material';
-import { ResultSnackbar } from '../../shared/dialogs/result-snackbar/result.snackbar';
+import { ResultSnackbar } from '@shared/dialogs/result-snackbar/result.snackbar';
 
-import { PersonType } from '../../shared/models/user.model';
-import { DocumentFile } from '../../shared/models/shared.model';
+import { PersonType } from '@shared/models/user.model';
+import { DocumentFile } from '@shared/models/shared.model';
 
 @Component({
   selector: 'app-signup',
@@ -24,6 +24,8 @@ export class SignupComponent implements OnInit {
 
   PersonType = PersonType;
   willRent = false;
+
+  terms = false;
 
   errors: any = {};
   userForm: FormGroup;
@@ -49,9 +51,8 @@ export class SignupComponent implements OnInit {
   get cRlIdentification() { return this.company.get('rl_identification'); }
 
   get renterForm() { return this.userForm.get('renter'); }
-  get matInmobiliaria() { return this.renterForm.get('matricula_inmobiliaria'); }
+  // get matInmobiliaria() { return this.renterForm.get('matricula_inmobiliaria'); }
   bankReference: DocumentFile;
-  certLibTra: DocumentFile;
   rut: DocumentFile;
 
   constructor(
@@ -64,31 +65,31 @@ export class SignupComponent implements OnInit {
 
   ngOnInit() {
     this.userForm = this.formBuilder.group({
-      email: ['prueba@prueba.com', [Validators.required, CustomValidators.email()]],
-      phone_number: ['3109878765', [Validators.minLength(7)]],
-      cell_phone: ['31052458547', [Validators.required, Validators.minLength(7)]],
+      email: ['', [Validators.required, CustomValidators.email()]],
+      phone_number: ['', [Validators.minLength(7)]],
+      cell_phone: ['', [Validators.required, Validators.minLength(7)]],
       passwords: this.formBuilder.group({
-        password: ['12345678', [Validators.required, Validators.minLength(8)]],
-        password_confirmation: ['12345678', [Validators.required, Validators.minLength(8)]],
+        password: ['', [Validators.required, Validators.minLength(8)]],
+        password_confirmation: ['', [Validators.required, Validators.minLength(8)]],
       }, { validator: [CustomValidators.matchPasswords] }),
 
       person: this.formBuilder.group({
-        name: ['Nombre Prueba', [Validators.required]],
-        last_name: ['Apellido Prueba', [Validators.required]],
-        identification: ['1140890987', [Validators.required, Validators.minLength(6)]],
+        name: ['', [Validators.required]],
+        last_name: ['', [Validators.required]],
+        identification: ['', [Validators.required, Validators.minLength(6)]],
       }),
 
       company: this.formBuilder.group({
-        name: ['Empresa de Prueba', [Validators.required]],
-        nit: ['800000000', [Validators.required]],
-        rl_name: ['800000000', [Validators.required]],
-        rl_last_name: ['800000000', [Validators.required]],
-        rl_identification: ['800000000', [Validators.required, Validators.minLength(6)]],
+        name: ['', [Validators.required]],
+        nit: ['', [Validators.required]],
+        rl_name: ['', [Validators.required]],
+        rl_last_name: ['', [Validators.required]],
+        rl_identification: ['', [Validators.required, Validators.minLength(6)]],
       }),
       personType: [PersonType.NATURAL, [Validators.required]],
 
       renter: this.formBuilder.group({
-        matricula_inmobiliaria: ['8asdsd4a6s5f4qaw896gf5q', [Validators.required]]
+        // matricula_inmobiliaria: ['', [Validators.required]]
       })
     });
   }
@@ -137,9 +138,8 @@ export class SignupComponent implements OnInit {
       const child = requestParams.company || requestParams.person;
       child.renter_attributes = {
         bank_reference: this.bankReference,
-        certificado_libertad_tradicion: this.certLibTra,
         rut: this.rut,
-        matricula_inmobiliaria: this.matInmobiliaria.value,
+        // matricula_inmobiliaria: this.matInmobiliaria.value,
       };
     }
     this.authService.signup(requestParams)
@@ -176,9 +176,6 @@ export class SignupComponent implements OnInit {
       case 0:
         this.bankReference = f;
         break;
-      case 1:
-        this.certLibTra = f;
-        break;
       case 2:
         this.rut = f;
         break;
@@ -198,16 +195,12 @@ export class SignupComponent implements OnInit {
     if (!this.willRent) {
       return false;
     }
-    return this.renterForm.valid &&
-      this.bankReference &&
-      this.certLibTra &&
-      this.rut;
+    return !(this.renterForm.valid && this.bankReference && this.rut);
   }
 
   onWillRentchange(accordion) {
     this.renterForm.reset();
     this.bankReference = undefined;
-    this.certLibTra = undefined;
     this.rut = undefined;
     this.willRent ? accordion.open() : accordion.close();
     console.log(this.willRent);
