@@ -1,12 +1,8 @@
-import { NgModule } from '@angular/core';
+import { NgModule, Optional, SkipSelf } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { HTTP_INTERCEPTORS } from '@angular/common/http';
 import { HttpClientModule } from '@angular/common/http';
-
-import { SharedService } from './providers/shared.service';
-import { AuthService } from './providers/auth.service';
-import { ModalManager } from './providers/modal-manager';
 
 import { ValidAuthInterceptor } from './interceptors/valid-auth.interceptor';
 import { AuthHeadersInterceptor } from './interceptors/auth-headers.interceptor';
@@ -14,12 +10,14 @@ import { AuthHeadersInterceptor } from './interceptors/auth-headers.interceptor'
 /*Overlays */
 import { OverlayModule } from '@angular/cdk/overlay';
 import { PdfViewerOverlayService } from '@shared/overlays/file-preview/pdf-viewer-overlay.service';
+import { MAT_STEPPER_GLOBAL_OPTIONS } from '@angular/cdk/stepper';
+
 
 @NgModule({
   imports: [
     CommonModule,
     HttpClientModule,
-    OverlayModule
+    OverlayModule,
     //
   ],
   providers: [
@@ -34,11 +32,22 @@ import { PdfViewerOverlayService } from '@shared/overlays/file-preview/pdf-viewe
       multi: true
     },
     //
-    SharedService,
-    AuthService,
-    ModalManager,
+    {
+      provide: MAT_STEPPER_GLOBAL_OPTIONS,
+      useValue: {
+        displayDefaultIndicatorType: false,
+        showError: true
+      }
+    },
     //
     PdfViewerOverlayService,
   ]
 })
-export class CoreModule { }
+export class CoreModule {
+  constructor(@Optional() @SkipSelf() parentModule: CoreModule) {
+    if (parentModule) {
+      throw new Error(
+        'CoreModule is already loaded. Import it in the AppModule only');
+    }
+  }
+}
