@@ -1,15 +1,15 @@
 import { NgModule } from '@angular/core';
-import { SharedModule } from '../../app/shared/shared.module';
+import { SharedModule } from '@shared/shared.module';
+import { UsedMaterialComponentsModule } from '@shared/modules/used-material-components.module';
 
 import { HomeComponent } from './home.component';
-
-// import { HomeService } from './home.service';
 
 import { HomeRoutingModule } from './home.routing';
 
 @NgModule({
   imports: [
     SharedModule,
+    UsedMaterialComponentsModule,
     HomeRoutingModule
   ],
   declarations: [
