@@ -1,2 +1,1 @@
-
-export * from './app-constants';
+export * from '@app/app-constants';
