@@ -16,13 +16,13 @@ import { Pipe, PipeTransform } from '@angular/core';
 export class CopCurrencyPipe implements PipeTransform {
   transform(value: number, showSymbol: boolean = true): string {
     let isNeg = false;
-    let val = '' + value;
+    const val = '' + value;
     if (value < 0) {
       val.substring(1);
       isNeg = true;
     }
-    let cop = ''
-    let times = Math.floor((val.length - 1) / 3);
+    let cop = '';
+    const times = Math.floor((val.length - 1) / 3);
     for (let i = 0; i < times; i++) {
       cop = '.' + (val + '.').slice(-4 - 3 * i, -1 - 3 * i) + cop;
     }
