@@ -14,14 +14,14 @@ import { Pipe, PipeTransform } from '@angular/core';
   name: 'date'
 })
 export class DatePipe implements PipeTransform {
-  transform(value: string) : string {
-    let date = new Date(value);
+  transform(value: string): string {
+    const date = new Date(value);
 
     let d = '' + ('0' + date.getDate()).slice(-2) + '/';
     d += ('0' + (date.getMonth() + 1)).slice(-2) + '/';
     d += date.getFullYear() + ' ';
 
-    let h = date.getHours();
+    const h = date.getHours();
     d += ('0' + (h <= 12 ? h : h - 12)).slice(-2) + ':';
 
     d += ('0' + date.getMinutes()).slice(-2) + ':';
