@@ -15,7 +15,7 @@ import { Pipe, PipeTransform } from '@angular/core';
   name: 'truncate'
 })
 export class TruncatePipe implements PipeTransform {
-  transform(value: string, limit: number = 10, trail: string = '...') : string {
+  transform(value: string, limit: number = 10, trail: string = '...'): string {
     return value ? value.length > limit ? value.substring(0, limit) + trail : value : '';
   }
 }
