@@ -1,4 +1,5 @@
 export const environment = {
   production: true,
-  GOOGLE_API_KEY: 'AIzaSyB_TNfE3rG_JOV-3jrDNT04lml2sxPzNUc'
+  GOOGLE_API_KEY: '',
+  API_ENDPOINT: '',
 };
