@@ -1,9 +1,18 @@
-// The file contents for the current environment will overwrite these during build.
-// The build system defaults to the dev environment which uses `environment.ts`, but if you do
-// `ng build --env=prod` then `environment.prod.ts` will be used instead.
-// The list of which env maps to which file can be found in `.angular-cli.json`.
+// This file can be replaced during build by using the `fileReplacements` array.
+// `ng build ---prod` replaces `environment.ts` with `environment.prod.ts`.
+// The list of file replacements can be found in `angular.json`.
 
 export const environment = {
   production: false,
-  GOOGLE_API_KEY: 'AIzaSyB_TNfE3rG_JOV-3jrDNT04lml2sxPzNUc',
+  GOOGLE_API_KEY: 'AIzaSyBTXShFoqSbulYFTQIr2_zD8Op4Ojyi-2g',
+  API_ENDPOINT: 'https://company-tonmil931210.c9users.io/',
+
 };
+
+/*
+ * In development mode, to ignore zone related error stack frames such as
+ * `zone.run`, `zoneDelegate.invokeTask` for easier debugging, you can
+ * import the following file, but please comment it out in production mode
+ * because it will have performance impact when throw error
+ */
+// import 'zone.js/dist/zone-error';  // Included with Angular CLI.
